+++ src/main/java/net/silentchaos512/gear/api/item/ICoreArmor.java
@@ -48,7 +48,7 @@ public interface ICoreArmor extends ICoreItem {
     }
 
     @Override
-    default boolean supportsPart(PartData part) {
+    default boolean supportsPart(ItemStack gear, PartData part) {
         PartType type = part.getType();
         return type == PartType.MAIN || type == PartType.TIP;
     }