+++ src/main/java/net/silentchaos512/gear/api/parts/IUpgradePart.java
@@ -6,6 +6,7 @@ import net.silentchaos512.gear.api.item.ICoreItem;
 /**
  * Represents an upgrade part, which is applied to gear after it has been crafted.
  */
+@Deprecated
 public interface IUpgradePart {
     /**
      * Check if the part can be applied to the gear type. This is <em>not</em> responsible for