+++ src/main/java/net/silentchaos512/gear/parts/type/BindingPart.java
@@ -18,12 +18,16 @@
 
 package net.silentchaos512.gear.parts.type;
 
+import net.minecraft.item.ItemStack;
 import net.minecraft.util.ResourceLocation;
+import net.silentchaos512.gear.api.item.GearType;
 import net.silentchaos512.gear.api.item.ICoreItem;
 import net.silentchaos512.gear.api.item.ICoreTool;
 import net.silentchaos512.gear.api.parts.*;
 import net.silentchaos512.gear.parts.AbstractGearPart;
+import net.silentchaos512.gear.parts.PartData;
 import net.silentchaos512.gear.parts.PartPositions;
+import net.silentchaos512.gear.util.GearHelper;
 
 public class BindingPart extends AbstractGearPart implements IUpgradePart {
     public BindingPart(ResourceLocation location) {
@@ -43,6 +47,12 @@ public class BindingPart extends AbstractGearPart implements IUpgradePart {
     @Override
     public IPartSerializer<?> getSerializer() {
         return PartType.BINDING.getSerializer();
+    }
+
+    @Override
+    public boolean canAddToGear(ItemStack gear, PartData part) {
+        GearType type = GearHelper.getType(gear);
+        return type != null && type.matches(GearType.TOOL);
     }
 
     @Override